+++ qatrack/qatrack_core/forms.py
@@ -1,3 +1,4 @@
+import json
 from django import forms
 
 
@@ -11,3 +12,12 @@ class MultipleCharField(forms.CharField):
 class UserChoiceField(forms.ModelChoiceField):
     def label_from_instance(self, obj):
         return obj.get_full_name()
+
+
+class JSONWidget(forms.widgets.Textarea):
+    """To be used in conjunction with qatrack.core.fields.JSONField"""
+
+    def format_value(self, value):
+        """Pretty print the Python object ensuring it is stringified in valid
+        JSON format"""
+        return json.dumps(value, indent=2, sort_keys=True)